move test, doesnt run in non-workspace crate
