+++ crates/template_builtin/templates/liquidity_pool/Cargo.toml
@@ -16,7 +16,3 @@ lto = true          # Enable Link Time Optimization.
 codegen-units = 1   # Reduce number of codegen units to increase optimizations.
 panic = 'abort'     # Abort on panic (reduces binary size).
 strip = 'debuginfo' # Strip debug info to reduce binary size.
-
-[dev-dependencies]
-tari_template_test_tooling = { path = "../../../template_test_tooling" }
-tari_transaction = { path = "../../../transaction" }
